backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
httpx==0.27.2
redis==5.1.1
websockets==13.1
apscheduler==3.10.4
greenlet==3.1.1
psycopg2-binary==2.9.9
email-validator==2.2.0
